# Meizon-Inventory-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Inventory Management Service",
    "version": "1.0.1",
    "title": "Meizon Inventory Service",
    "license": {
      "name": "Apache Licence Version 2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/daily/readings/user/{userId}": {
      "get": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "getDailyReadingById",
        "operationId": "getDailyReadingByIdUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/user/": {
      "get": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "getAllUser",
        "operationId": "getAllUserUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "saveUser",
        "operationId": "saveUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDto",
            "description": "userDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/user/{locationId}": {
      "get": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "getUserByLocationId",
        "operationId": "getUserByLocationIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/user/{roleId}": {
      "get": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "getUserByRoleId",
        "operationId": "getUserByRoleIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "roleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/user/{userId}": {
      "get": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "getUserByUserId",
        "operationId": "getUserByUserIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "editUser",
        "operationId": "editUserUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDto",
            "description": "userDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "user-controller-v-1"
        ],
        "summary": "deletUserByID",
        "operationId": "deletUserByIDUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "AddressDto": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string"
        },
        "unitNumber": {
          "type": "string"
        }
      },
      "title": "AddressDto"
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "locationId": {
          "type": "integer",
          "format": "int64"
        },
        "mobile": {
          "type": "integer",
          "format": "int64"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserRoleDto"
          }
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "UserDto"
    },
    "UserPermissionDto": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "UserPermissionDto"
    },
    "UserRoleDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "permissions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserPermissionDto"
          }
        }
      },
      "title": "UserRoleDto"
    }
  }
}